Suspend post_save receivers during test v4: solve cyclic imports
