+++ tardis/tardis_portal/models/hooks.py
@@ -54,15 +54,3 @@ def post_save_experiment_parameter(sender, **kwargs):
 def post_save_experiment(sender, **kwargs):
     experiment = kwargs['instance']
     publish_public_expt_rifcs(experiment)
-
-
-def suspendingreceiver(signal, **decorator_kwargs):
-    def our_wrapper(func):
-        @receiver(signal, **decorator_kwargs)
-        @functools.wraps(func)
-        def fake_receiver(sender, **kwargs):
-            if settings.SUSPEND_SIGNALS:
-                return
-            return func(sender, **kwargs)
-        return fake_receiver
-    return our_wrapper
